renamed tx_map module
