+++ zingolib/src/wallet/tx_map/trait_stub_inputsource.rs
