+++ zingolib/src/wallet/tx_map/trait_walletread.rs
